+++ typings/tests/styled-components-test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
-import styled from "..";
-import { css, keyframes, ThemeProvider, injectGlobal, withTheme, ServerStyleSheet } from "..";
+import styled from "../..";
+import { css, keyframes, ThemeProvider, injectGlobal, withTheme, ServerStyleSheet } from "../..";
 
 // Create a <Title> react component that renders an <h1> which is
 // centered, palevioletred and sized at 1.5em
