+++ typings/tests/themed-tests/mytheme-styled-components-test.tsx
